Moved ComputeNode to its own file.
