+++ lluvia/cpp/core/include/lluvia/core.h
@@ -29,15 +29,15 @@
 #include "core/image/ImageView.h"
 #include "core/image/ImageViewDescriptor.h"
 
+#include "core/node/ComputeNode.h"
+#include "core/node/Node.h"
 #include "core/node/NodeState.h"
 #include "core/node/NodeType.h"
+#include "core/node/PortDescriptor.h"
 #include "core/node/PortDirection.h"
 #include "core/node/PortType.h"
-#include "core/node/PortDescriptor.h"
-#include "core/node/Node.h"
 
 #include "core/CommandBuffer.h"
-#include "core/ComputeNode.h"
 #include "core/ComputeNodeDescriptor.h"
 #include "core/ContainerNode.h"
 #include "core/ContainerNodeDescriptor.h"
