+++ lluvia/cpp/core/src/Session.cpp
@@ -11,7 +11,7 @@
 
 #include "lluvia/core/buffer/Buffer.h"
 #include "lluvia/core/CommandBuffer.h"
-#include "lluvia/core/ComputeNode.h"
+#include "lluvia/core/node/ComputeNode.h"
 #include "lluvia/core/ComputeNodeDescriptor.h"
 #include "lluvia/core/ContainerNode.h"
 #include "lluvia/core/ContainerNodeDescriptor.h"
