+++ lluvia/python/src/lluvia/core/node_new/__init__.py
@@ -6,8 +6,9 @@
     :license: Apache-2 license, see LICENSE for more details.
 """
 
+from lluvia.core.node_new.compute_node import *
 from lluvia.core.node_new.node_state import *
 from lluvia.core.node_new.node_type import *
+from lluvia.core.node_new.port_descriptor import *
 from lluvia.core.node_new.port_direction import *
 from lluvia.core.node_new.port_type import *
-from lluvia.core.node_new.port_descriptor import *
